Show date + add an option to ask a certain amount instead of all or nothing when showing search result (on !search/!last)
