+++ discord/botToken.json
@@ -1,0 +1,4 @@
+{
+  "dev": "your id",
+  "prod": "your id"
+}
