+++ discord/lib/Db/DbAccess.js
@@ -1,0 +1,1 @@
+module.exports = require('rethinkdbdash')({db: "RcArchiver"});
