+++ discord/lib/Driver/IRcDriver.js
@@ -1,0 +1,7 @@
+// So fucking useless, but let me mind like if this was an interface.
+class IRcDriver {
+    static GetRcData(msg){}
+    static GetLink(msg){}
+}
+
+module.exports = IRcDriver;
