+++ package.json
@@ -1,15 +1,18 @@
 {
   "name": "rc-archiver",
-  "version": "1.0.3",
+  "version": "1.0.4",
   "dev": false,
   "dependencies": {
     "bluebird": "^3.5.0",
     "chalk": "^2.0.1",
     "dateformat": "^3.0.2",
-    "discord-akairo": "^7.4.2",
+    "discord-akairo": "^7.5.1",
     "discord.js": "^11.1.0",
+    "discord.js-commando": "^0.9.0",
     "es6-enum": "^1.1.0",
     "hashids": "^1.1.1",
+    "iconv": "latest",
+    "joi": "^13.0.2",
     "jsdom": "^11.1.0",
     "linq-es6": "^0.2.1",
     "load-grunt-tasks": "^3.5.2",
@@ -17,9 +20,12 @@
     "moment": "^2.18.1",
     "node-cleanup": "^2.1.2",
     "phantom": "^4.0.4",
+    "requelize": "^0.9.6",
     "request": "^2.82.0",
     "rethinkdbdash": "^2.3.29",
+    "sequelize": "^4.28.6",
     "shortid": "^2.2.8",
+    "sqlite": "^2.9.0",
     "time-grunt": "^1.4.0",
     "uuid": "^3.1.0",
     "vue": "^2.3.3",
@@ -30,9 +36,7 @@
   "author": "Maxime Bonvin <max.bonvin@gmail.com>",
   "private": true,
   "scripts": {
-    "dev": "node build/dev-server.js",
-    "start": "node build/dev-server.js",
-    "build": "node build/build.js"
+    "dev": "node discord/bot.js"
   },
   "devDependencies": {
     "autoprefixer": "^7.1.2",
