--- web/config/prod.env.js
@@ -1,3 +1,0 @@
-module.exports = {
-  NODE_ENV: '"production"'
-}
