--- web/src/components/About.vue
@@ -1,20 +1,0 @@
-<template>
-  <div class="about">
-    <h1>it's about :D</h1>
-  </div>
-</template>
-
-<script>
-import Vue from "vue";
-
-Vue.material.setCurrentTheme('app');
-
-export default {
-  name: 'about',
-  data () {
-    return {
-
-    }
-  }
-}
-</script>
